requirements.txt
psutil==5.9.5
pynput==1.7.6
python_nmap==0.7.1
pywin32==305
Requests==2.31.0
scapy==2.4.5
speedtest==0.0.1
speedtest_cli==2.1.3
ttkbootstrap==1.10.1